// apps/playground-flow-refactor/src/app/components/workflow/types.ts
export enum BlockEnum {
  Start = "start",
  End = "end",
  Answer = "answer",
  LLM = "llm",
  KnowledgeRetrieval = "knowledge-retrieval",
  QuestionClassifier = "question-classifier",
  IfElse = "if-else",
  Code = "code",
  TemplateTransform = "template-transform",
  HttpRequest = "http-request",
  VariableAssigner = "variable-assigner",
  VariableAggregator = "variable-aggregator",
  Tool = "tool",
  ParameterExtractor = "parameter-extractor",
  Iteration = "iteration",
  DocExtractor = "document-extractor",
  ListFilter = "list-operator",
  IterationStart = "iteration-start",
  Assigner = "assigner", // is now named as VariableAssigner
  Agent = "agent",
  Loop = "loop",
  LoopStart = "loop-start",
  LoopEnd = "loop-end",
}
// ...
export type ValueSelector = string[];
// ...
export enum VarType {
  string = "string",
  number = "number",
  secret = "secret",
  boolean = "boolean",
  object = "object",
  file = "file",
  array = "array",
  arrayString = "array[string]",
  arrayNumber = "array[number]",
  arrayObject = "array[object]",
  arrayFile = "array[file]",
  any = "any",
}

export enum ValueType {
  variable = "variable",
  constant = "constant",
}

export type Var = {
  variable: string;
  type: VarType;
  children?: Var[] | StructuredOutput; // if type is obj, has the children struct
  isParagraph?: boolean;
  isSelect?: boolean;
  options?: string[];
  required?: boolean;
  des?: string;
  isException?: boolean;
  isLoopVariable?: boolean;
  nodeId?: string;
};